fix: QoL improvements for migration scripts
* Rename to the more appropriate `generate`, as we make database changes
  now instead of just outputting JSON
* Require a `--confirm` argument before dropping any collections
* Clean-up connections after encountering errors during migration
