+++ packages/migrate/package.json
@@ -8,8 +8,8 @@
     "lodash": "4.17.10"
   },
   "scripts": {
-    "export-new-reports": "node ./bin/export-new-reports.js > ./data/new-reports.json",
-    "export-new-loos": "node ./bin/export-new-loos.js > ./data/new-loos.json",
+    "generate-new-reports": "node ./bin/generate-new-reports.js > ./data/new-reports.json",
+    "generate-new-loos": "node ./bin/generate-new-loos.js > ./data/new-loos.json",
     "summarise-report-values": "node ./bin/summarise-report-values.js < ./data/new-reports.json > ./data/report-values.md"
   }
 }
